Update solution - CodeHub

+++ 7-kyu/most-digits/most-digits.java
@@ -1,7 +1,8 @@
+ 
 public class MostDigits {
   public static int findLongest(int[] numbers) {
     int result =0;
-    for(int i=0; i<numbers.length; i++){
+    for(long i=0; i<numbers.length; i++){
       if(result<numbers[i]){
         result = numbers[i];
       }
